src/web-backend/tsan_server/requirements.txt
django
graphene-django
djangorestframework
djangorestframework-jwt
django-cors-headers
django-admin-rangefilter
pymongo